fix(shard): drop two no-op u64 conversions
`RETENTION_EPOCHS` and `EPOCH_BLOCKS.get()` are already u64, so the `u64::from`
wrappers are `clippy::useless_conversion` and CI runs clippy with -D warnings.

Caught by CI, not locally, because I ran `cargo clippy -p ghost-pool --tests`
while CI runs `--workspace --all-targets --all-features`. That is the second time
today a narrower local clippy passed something the gate rejects — the first was
`--tests` being omitted. Reproduce the CI invocation, not an approximation of it.

+++ bins/ghost-pool/src/shard.rs
@@ -857,7 +857,7 @@ impl ShardRuntime {
         // `latest - RETENTION_EPOCHS`, so including it picks an epoch whose leaves are already
         // gone — the responder rebuilds a short tree, the root check fails, and the audit burns a
         // pending slot for an hour learning nothing.
-        let oldest = latest.saturating_sub(u64::from(RETENTION_EPOCHS).saturating_sub(1));
+        let oldest = latest.saturating_sub(RETENTION_EPOCHS.saturating_sub(1));
         let mut candidates = Vec::new();
         for epoch in oldest..=latest {
             if let Some(summary) = self.db.shard_get_epoch(epoch, target)? {
